Make it possible to use an ecosystem describer in more cases
Before this commit the describer would only be used if the same set of attributes
was found. This means that if the consumer added, or removed, one attribute, we
would lose the benefit of better user error messages. With this change, we try
to find the _best matching_ describer, if any.